requirements.txt
# Core dependencies
crewai[tools]==0.140.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database and storage
chromadb==0.5.23
sqlalchemy==2.0.36
alembic==1.14.0

# Memory and knowledge
mem0ai==0.1.114

# Monitoring and logging
structlog==24.4.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
pyyaml==6.0.2